// src/views/arap/payable/interfaces/PayableGrid.ts
import { Filter, PagesInfo } from "@/views/arap/_commons/interfaces/Grid";
import { ReactNode } from "react";


export interface PayableGrid {
  payableGridList: PayableColumn[];
  isLoading: boolean;
  filters: Filter[];
  info: PagesInfo;
}

export interface PayableColumn {
  account_id: number;
  capability_id: number;
  capability_name: string;
  currency: string;
  "120": number | null;
  "0_30": number | null;
  "31_60": number | null;
  "61_90": number | null;
  "91_120": number | null;
  total_debt: number;
  paid_percent: number;
}


export interface PayableCollapsibleData {
  collapsibleRow: ReactNode;
  collapsibleRowData: PayableColumn[];
}